Refactor javacsript into dedicated modules
