+++ src/web/index.html
@@ -35,7 +35,6 @@
     </div>
 </div>
 
-<script type="module" src="../../pkg/wiggers_graaf.js"></script>
-<script type="module" src="script.js"></script>
+<script type="module" src="scripts/main.js"></script>
 </body>
 </html>
